Move tests to separate class, small fixes
