+++ src/DI.Abstractions/Resources.resx
@@ -142,7 +142,7 @@
     <comment>{0} = implementation type, {1} = service type</comment>
   </data>
   <data name="WrongImplemetationType" xml:space="preserve">
-    <value>Implementation type '{0}' cann't be assigned to service type '{1}'.</value>
+    <value>Implementation type '{0}' can't be assigned to service type '{1}'.</value>
     <comment>{0} = implementation type, {1} = service type</comment>
   </data>
 </root>
